# restapi/src/main/webapp/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Provys Catalogue repository Api
  description: Provides methods for accessing entities and other objects from provys
    metadata catalogue
  version: "1.0"
servers:
- url: /api
  variables: {}
paths:
  /attr/{attrId}:
    get:
      summary: Get Attribute
      description: Retrieve atribute using Id
      operationId: getAttrById
      parameters:
      - name: attrId
        in: path
        required: true
        schema:
          type: integer
      responses:
        default:
          description: Attribute
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Attr'
  /attr/{entityNm}/{nameNm}:
    get:
      summary: Get Attribute by Internal Name
      description: Retrieve attribute using internal name
      operationId: getAttrByNameNm
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      - name: nameNm
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: Attribute
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Attr'
  /domain/{domainId}:
    get:
      summary: Get Domain
      description: Retrieve domain using Id
      operationId: getDomainById
      parameters:
      - name: domainId
        in: path
        required: true
        schema:
          type: integer
      responses:
        default:
          description: default response
          content:
            application/json: {}
            application/xml: {}
  /domain/{domainNm}:
    get:
      summary: Get Domain by Internal Name
      description: Retrieve domain using internal name
      operationId: getDomainByNameNm
      parameters:
      - name: domainNm
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
            application/xml:
              schema:
                $ref: '#/components/schemas/Domain'
  /domain:
    get:
      summary: Get All Domains
      description: Retrieve all domains
      operationId: getAllDomains
      responses:
        default:
          description: default response
          content:
            application/json: {}
            application/xml: {}
  /entity/{entityNm}:
    get:
      summary: Get Entity by Internal Name
      description: Retrieve entity using internal name
      operationId: getEntityByNameNm
      parameters:
      - name: entityNm
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
  /entity/{entityId}:
    get:
      summary: Get Entity
      description: Retrieve entity using Id
      operationId: getEntityByNameNm_1
      parameters:
      - name: entityId
        in: path
        required: true
        schema:
          type: integer
      responses:
        default:
          description: Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
components:
  schemas:
    Attr:
      type: object
      properties:
        name:
          type: string
        entity:
          $ref: '#/components/schemas/Entity'
        attrGrpId:
          type: integer
        domain:
          $ref: '#/components/schemas/Domain'
        subdomainNm:
          type: string
        mandatory:
          type: boolean
        ordInEntity:
          type: integer
          format: int32
        nameNm:
          type: string
        note:
          type: string
        properNameRoot:
          type: string
        attrGrp:
          $ref: '#/components/schemas/AttrGrp'
        entityId:
          type: integer
        ord:
          type: integer
          format: int32
        attrType:
          type: string
          enum:
          - COLUMN
          - TEXT
          - FREE
          - FINANCE
          - SQL
          - FORMULA
        defValue:
          type: string
        javaName:
          type: string
        id:
          type: integer
    AttrGrp:
      type: object
      properties:
        name:
          type: string
        entity:
          $ref: '#/components/schemas/Entity'
        nameNm:
          type: string
        note:
          type: string
        entityId:
          type: integer
        ord:
          type: integer
          format: int32
        attrs:
          type: array
          items:
            $ref: '#/components/schemas/Attr'
        id:
          type: integer
    Domain:
      type: object
      properties:
        name:
          type: string
          xml:
            name: NAME
        note:
          type: string
        validateCd:
          type: string
        natOrder:
          type: boolean
        allowed:
          type: boolean
        isqVisible:
          type: boolean
        getlFormatCd:
          type: string
        islVisible:
          type: boolean
        dataPrecision:
          type: integer
          format: int32
        dataTypeNm:
          type: string
        dataLength:
          type: integer
          format: int32
        nameNm:
          type: string
        id:
          type: integer
    Entity:
      type: object
      properties:
        name:
          type: string
        table:
          type: string
        ancestor:
          $ref: '#/components/schemas/Entity'
        note:
          type: string
        custom:
          type: boolean
        used:
          type: boolean
        objectClass:
          type: boolean
        pgPackage:
          type: string
        ancestorId:
          type: integer
        epPackage:
          type: string
        entityGrp:
          $ref: '#/components/schemas/EntityGrp'
        structureDoc:
          type: string
        usageDoc:
          type: string
        cpPackage:
          type: string
        customNote:
          type: string
        view:
          type: string
        fpPackage:
          type: string
        entityGrpId:
          type: integer
        behaviourDoc:
          type: string
        implDoc:
          type: string
        attrGrps:
          type: array
          items:
            $ref: '#/components/schemas/AttrGrp'
        attrs:
          type: array
          items:
            $ref: '#/components/schemas/Attr'
        nameNm:
          type: string
        id:
          type: integer
    EntityGrp:
      type: object
      properties:
        name:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/EntityGrp'
        note:
          type: string
        ord:
          type: integer
          format: int32
        parentId:
          type: integer
        fullOrd:
          type: array
          items:
            type: integer
            format: int32
        entities:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        nameNm:
          type: string
        id:
          type: integer